WIP adding method support to SampleGen
