+++ CommunityToolkit.Tooling.SampleGen.Tests/ToolkitSampleMetadataTests.cs
@@ -16,342 +16,6 @@ namespace CommunityToolkit.Tooling.SampleGen.Tests;
 public partial class ToolkitSampleMetadataTests
 {
     [TestMethod]
-    public void PaneOption_GeneratesWithoutDiagnostics()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""Test"", false, Title = ""Toggle y"")]
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", ""Segoe UI"", ""Arial"", ""Consolas"", Title = ""Font family"")]
-
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                    public Sample()
-                    {{
-                        var x = this.Test;
-                        var y = this.TextFontFamily;
-                    }}
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleOptionGenerator>(source, string.Empty);
-    }
-
-    [TestMethod]
-    public void PaneOption_GeneratesTitleProperty()
-    {
-        var source = """
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {
-                [ToolkitSampleNumericOption("TextSize", 12, 8, 48, 2, false, Title = "FontSize")]
-                [ToolkitSample(id: nameof(Sample), "Test Sample", description: "")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {
-                    public Sample()
-                    {
-                        var x = this.Test;
-                        var y = this.TextFontFamily;
-                    }
-                }
-            }
-
-            namespace Windows.UI.Xaml.Controls
-            {
-                public class UserControl { }
-            }
-        """;
-
-        var result = """
-            #nullable enable
-            namespace CommunityToolkit.Tooling.SampleGen;
-
-            public static class ToolkitSampleRegistry
-            {
-                public static System.Collections.Generic.Dictionary<string, CommunityToolkit.Tooling.SampleGen.Metadata.ToolkitSampleMetadata> Listing
-                { get; } = new() {
-                    ["Sample"] = new CommunityToolkit.Tooling.SampleGen.Metadata.ToolkitSampleMetadata("Sample", "Test Sample", "", typeof(MyApp.Sample), () => new MyApp.Sample(), null, null, new CommunityToolkit.Tooling.SampleGen.Metadata.IGeneratedToolkitSampleOptionViewModel[] { new CommunityToolkit.Tooling.SampleGen.Metadata.ToolkitSampleNumericOptionMetadataViewModel(name: "TextSize", initial: 12, min: 8, max: 48, step: 2, showAsNumberBox: false, title: "FontSize") })
-                };
-            }
-            """;
-
-        VerifyGenerateSources("MyApp.Tests", source, new[] { new ToolkitSampleMetadataGenerator() }, ignoreDiagnostics: true, ("ToolkitSampleRegistry.g.cs", result));
-    }
-
-    // https://github.com/CommunityToolkit/Labs-Windows/issues/175
-    [TestMethod]
-    public void PaneOption_GeneratesProperty_DuplicatePropNamesAcrossSampleClasses()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""Test"", false, Title = ""Toggle y"")]
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", ""Segoe UI"", ""Arial"", ""Consolas"", Title = ""Font family"")]
-
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                    public Sample()
-                    {{
-                        var x = this.Test;
-                        var y = this.TextFontFamily;
-                    }}
-                }}
-
-                [ToolkitSampleBoolOption(""Test"", false, Title = ""Toggle y"")]
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", ""Segoe UI"", ""Arial"", ""Consolas"", Title = ""Font family"")]
-
-                [ToolkitSample(id: nameof(Sample2), ""Test Sample"", description: """")]
-                public partial class Sample2 : Windows.UI.Xaml.Controls.UserControl
-                {{
-                    public Sample2()
-                    {{
-                        var x = this.Test;
-                        var y = this.TextFontFamily;
-                    }}
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleOptionGenerator>(source, string.Empty);
-    }
-
-    [TestMethod]
-    public void PaneOptionOnNonSample()
-    {
-        string source = @"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {
-                [ToolkitSampleBoolOption(""BindToMe"", false, Title =  ""Toggle visibility"")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {
-                }
-            }
-
-            namespace Windows.UI.Xaml.Controls
-            {
-                public class UserControl { }
-            }";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneOptionAttributeOnNonSample.Id);
-    }
-
-    [DataRow("", DisplayName = "Empty string"), DataRow(" ", DisplayName = "Only whitespace"), DataRow("Test ", DisplayName = "Text with whitespace")]
-    [DataRow("_", DisplayName = "Underscore"), DataRow("$", DisplayName = "Dollar sign"), DataRow("%", DisplayName = "Percent symbol")]
-    [DataRow("class", DisplayName = "Reserved keyword 'class'"), DataRow("string", DisplayName = "Reserved keyword 'string'"), DataRow("sealed", DisplayName = "Reserved keyword 'sealed'"), DataRow("ref", DisplayName = "Reserved keyword 'ref'")]
-    [TestMethod]
-    public void PaneOptionWithBadName(string name)
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                [ToolkitSampleBoolOption(""{name}"", false, Title =  ""Toggle visibility"")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneOptionWithBadName.Id, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void PaneOptionWithConflictingPropertyName()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""IsVisible"", false, Title =  ""Toggle x"")]
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                    public string IsVisible {{ get; set; }}
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneOptionWithConflictingName.Id, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void PaneOptionWithConflictingInheritedPropertyName()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""IsVisible"", false, Title =  ""Toggle x"")]
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Base
-                {{
-                }}
-
-                public class Base : Windows.UI.Xaml.Controls.UserControl
-                {{
-                    public string IsVisible {{ get; set; }}
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneOptionWithConflictingName.Id, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void PaneOptionWithDuplicateName()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""test"", false, Title =  ""Toggle x"")]
-                [ToolkitSampleBoolOption(""test"", false, Title =  ""Toggle y"")]
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", ""Segoe UI"", ""Arial"", Title = ""Text foreground"")]
-
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneOptionWithDuplicateName.Id, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void PaneOptionWithDuplicateName_AllowedBetweenSamples()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleBoolOption(""test"", false, Title =  ""Toggle y"")]
-
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                }}
-
-                [ToolkitSampleBoolOption(""test"", false, Title =  ""Toggle y"")]
-
-                [ToolkitSample(id: nameof(Sample2), ""Test Sample"", description: """")]
-                public partial class Sample2 : Windows.UI.Xaml.Controls.UserControl
-                {{
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void PaneMultipleChoiceOptionWithNoChoices()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", Title = ""Font family"")]
-
-                [ToolkitSample(id: nameof(Sample), ""Test Sample"", description: """")]
-                public partial class Sample : Windows.UI.Xaml.Controls.UserControl
-                {{
-                }}
-            }}
-
-            namespace Windows.UI.Xaml.Controls
-            {{
-                public class UserControl {{ }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SamplePaneMultiChoiceOptionWithNoChoices.Id, DiagnosticDescriptors.SampleNotReferencedInMarkdown.Id);
-    }
-
-    [TestMethod]
-    public void SampleGeneratedOptionAttributeOnUnsupportedType()
-    {
-        var source = $@"
-            using System.ComponentModel;
-            using CommunityToolkit.Tooling.SampleGen;
-            using CommunityToolkit.Tooling.SampleGen.Attributes;
-
-            namespace MyApp
-            {{
-                [ToolkitSampleMultiChoiceOption(""TextFontFamily"", ""Segoe UI"", ""Arial"", ""Consolas"", Title = ""Font family"")]
-                [ToolkitSampleBoolOption(""Test"", false, Title =  ""Toggle visibility"")]
-                public partial class Sample
-                {{
-                }}
-            }}";
-
-        VerifyGeneratedDiagnostics<ToolkitSampleMetadataGenerator>(source, string.Empty, DiagnosticDescriptors.SampleGeneratedOptionAttributeOnUnsupportedType.Id, DiagnosticDescriptors.SamplePaneOptionAttributeOnNonSample.Id);
-    }
-
-    [TestMethod]
     public void SampleAttributeOnUnsupportedType()
     {
         var source = $@"
@@ -432,7 +96,7 @@ public partial class ToolkitSampleMetadataTests
     /// <param name="source">The input source to process.</param>
     /// <param name="markdown">The input documentation info to process.</param>
     /// <param name="diagnosticsIds">The diagnostic ids to expect for the input source code.</param>
-    private static void VerifyGeneratedDiagnostics<TGenerator>(string source, string markdown, params string[] diagnosticsIds)
+    internal static void VerifyGeneratedDiagnostics<TGenerator>(string source, string markdown, params string[] diagnosticsIds)
         where TGenerator : class, IIncrementalGenerator, new()
     {
         VerifyGeneratedDiagnostics<TGenerator>(CSharpSyntaxTree.ParseText(source), markdown, diagnosticsIds);
@@ -445,7 +109,7 @@ public partial class ToolkitSampleMetadataTests
     /// <param name="syntaxTree">The input source tree to process.</param>
     /// <param name="markdown">The input documentation info to process.</param>
     /// <param name="diagnosticsIds">The diagnostic ids to expect for the input source code.</param>
-    private static void VerifyGeneratedDiagnostics<TGenerator>(SyntaxTree syntaxTree, string markdown, params string[] diagnosticsIds)
+    internal static void VerifyGeneratedDiagnostics<TGenerator>(SyntaxTree syntaxTree, string markdown, params string[] diagnosticsIds)
         where TGenerator : class, IIncrementalGenerator, new()
     {
         var sampleAttributeType = typeof(ToolkitSampleAttribute);
@@ -496,7 +160,7 @@ public partial class ToolkitSampleMetadataTests
     /// <param name="source">The input source to process.</param>
     /// <param name="generators">The generators to apply to the input syntax tree.</param>
     /// <param name="results">The source files to compare.</param>
-    private static void VerifyGenerateSources(string assemblyName, string source, IIncrementalGenerator[] generators, bool ignoreDiagnostics = false, params (string Filename, string Text)[] results)
+    internal static void VerifyGenerateSources(string assemblyName, string source, IIncrementalGenerator[] generators, bool ignoreDiagnostics = false, params (string Filename, string Text)[] results)
     {
         // Ensure our types are loaded
         Type sampleattributeObjectType = typeof(ToolkitSampleAttribute);
